pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "servicenow-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server implementation for ServiceNow"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "ServiceNow MCP Contributors"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "mcp[cli]==1.3.0",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.27.0",
    "uvicorn>=0.22.0",
    "httpx>=0.24.0",
    "PyYAML>=6.0",
    "beautifulsoup4>=4.12.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.0.1",
]

[project.scripts]
servicenow-mcp = "servicenow_mcp.cli:main"
servicenow-mcp-sse = "servicenow_mcp.server_sse:main"
moveworks-mcp = "moveworks_mcp.cli:main"
moveworks-mcp-sse = "moveworks_mcp.server_sse:main"

[tool.hatch.build.targets.wheel]
packages = ["src/servicenow_mcp", "src/moveworks_mcp"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--ignore=examples" 
